refactor: move math utils to separate file
